消息队列 大动脉

+++ proc/operating/exchange/spot/src/proc/behavior/spot_user_data_stream_behavior_template.rs
@@ -20,7 +20,7 @@ pub struct AccountCmd {
     /// 仅返回非零余额，默认 false
     pub omit_zero_balances: Option<bool>,
     /// 接收窗口（微秒精度），不超过 60000
-    pub recv_window: Option<f64>,
+    pub recv_window: Option<u64>,
     /// 时间戳
     pub timestamp: i64,
 }
